size_t is not an int, nor is it an unsigned long int, it's a size_t
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@6098 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/VotingBooth.h
@@ -84,13 +84,13 @@ protected:
   /** returns the number of voters that have participated
    */
   inline unsigned long int getVoterCount(void) const {
-    return _vote.size();
+    return (unsigned long int)_vote.size();
   }
 
   /** returns the number of responses available
    */
   inline unsigned long int getResponseCount(void) const {
-    return _choice.size();
+    return (unsigned long int)_choice.size();
   }
 
   /** returns a string identifier for this poll
